LPS-9568
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@52491 05bdf26c-840f-0410-9ced-eb539d925f36

+++ support-maven/plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/PluginDeployerMojo.java
@@ -39,11 +39,17 @@ public class PluginDeployerMojo extends AbstractMojo {
 	}
 
 	protected void doExecute() throws Exception {
-		getLog().info(
-			"Deploying " + warFileName + " to " +
+		if (warFile.exists()) {
+			getLog().info(
+				"Deploying " + warFileName + " to " +
 				autoDeployDir.getAbsolutePath());
 
-		FileUtils.copyFile(warFile, new File(autoDeployDir, warFileName));
+			FileUtils.copyFile(warFile, new File(autoDeployDir, warFileName));
+		}
+		else {
+			getLog().warn(
+				warFileName + " does not exist and won't be deployed.");
+		}
 	}
 
 	/**
